LPS-69710 Move impl classes to new module
