+++ modules/apps/foundation/microsoft-translator/microsoft-translator/src/main/java/com/liferay/microsoft/translator/internal/MicrosoftTranslatorAuthenticator.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.microsofttranslator;
+package com.liferay.microsoft.translator.internal;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
